pyproject.toml
[project]
name = "portia-digest-bot"
version = "0.2.0"
description = "Fetch Portia plans and plan-runs, analyze, summarize, and email a daily digest."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27",
  "typer>=0.12",
  "python-dotenv>=1.0",
  "pydantic>=2.0",
  "portia-sdk-python>=0.7.2",
  "rich>=13.0"
]
[project.scripts]
portia-fetch = "portia_fetch.cli:app"
[tool.setuptools.packages.find]
where = ["."]
include = ["portia_fetch*"]
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
